Move `DBTask` out of helper package
This class belongs to the outermost layer as it does a lot of operational
things (like setting a session) – but the `helpers` package is meant
to be the innermost one (i.e., not depending on database interna or
model stuff)
