+++ doc/api/pycroft.helpers.rst
@@ -8,11 +8,5 @@ pycroft.helpers
 .. automodule:: pycroft.helpers.i18n
 .. automodule:: pycroft.helpers.interval
 .. automodule:: pycroft.helpers.net
-.. automodule:: pycroft.helpers.task
-    :exclude-members: DBTask
-
-    .. autoclass:: pycroft.helpers.task.DBTask
-        :members:
-        :exclude-members: ignore_result,priority,rate_limit,reject_on_worker_lost,request_stack,serializer,track_started,store_errors_even_if_ignored,typing
 .. automodule:: pycroft.helpers.user
 .. automodule:: pycroft.helpers.utc
