+++ pycroft/task.py
@@ -9,18 +9,19 @@ by using implementations as defined in `pycroft.lib.task`
 """
 import logging
 import os
+import sys
+import typing as t
 from datetime import timedelta
 
 import sentry_sdk
-from celery import Celery
+from celery import Celery, Task as CeleryTask
 from celery.schedules import crontab
 from celery.utils.log import get_task_logger
 from sentry_sdk.integrations.celery import CeleryIntegration
 from sentry_sdk.integrations.logging import LoggingIntegration
-from sqlalchemy.orm import with_polymorphic
+from sqlalchemy.orm import with_polymorphic, Session, scoped_session, sessionmaker
 from sqlalchemy.orm.exc import ObjectDeletedError
 
-from pycroft.helpers.task import DBTask
 from pycroft.lib.finance import get_negative_members
 from pycroft.lib.logging import log_task_event
 from pycroft.lib.mail import send_mails, Mail, RetryableException, \
@@ -29,11 +30,11 @@ from pycroft.lib.mail import send_mails, Mail, RetryableException, \
 from pycroft.lib.task import get_task_implementation
 from pycroft.model import session
 from pycroft.model.finance import BankAccountActivity
-from pycroft.model.session import with_transaction
+from pycroft.model.session import with_transaction, session, set_scoped_session
 from pycroft.model.swdd import swdd_vo, swdd_import, swdd_vv
 from pycroft.model.task import Task, TaskStatus
 from pycroft.model.traffic import TrafficVolume
-
+from scripts.connection import try_create_connection
 
 if dsn := os.getenv('PYCROFT_SENTRY_DSN'):
     logging_integration = LoggingIntegration(
@@ -67,6 +68,53 @@ def repair_session():
     if not session.session.is_active:
         session.session.rollback()
         print("Repaired session (rollback).")
+
+
+class DBTask(CeleryTask):
+    """
+    Base class for tasks which use the database.
+    """
+
+    connection = None
+    engine = None
+
+    def run(self, *args: t.Any, **kwargs: t.Any) -> None:
+        pass
+
+    def after_return(
+        self,
+        status: str,
+        retval: t.Any,
+        task_id: str,
+        args: tuple,
+        kwargs: dict,
+        einfo: t.Any = None,
+    ) -> None:
+        session.close()
+
+    def __init__(self) -> None:
+        in_celery = sys.argv and sys.argv[0].endswith("celery") and "worker" in sys.argv
+        if not in_celery:
+            return
+        try:
+            connection_string = os.environ["PYCROFT_DB_URI"]
+        except KeyError:
+            raise RuntimeError(
+                "Environment variable PYCROFT_DB_URI must be "
+                "set to an SQLAlchemy connection string."
+            )
+
+        self.connection, self.engine = try_create_connection(
+            connection_string, 5, logger=logging.getLogger("tasks"), echo=False
+        )
+
+        set_scoped_session(
+            t.cast(Session, scoped_session(sessionmaker(bind=self.engine)))
+        )
+
+    def __del__(self) -> None:
+        if self.connection is not None:
+            self.connection.close()
 
 
 @app.task(base=DBTask)
